Update About section integration test suite
