+++ src/sections/About.js
@@ -41,7 +41,7 @@ function AboutTabs(props) {
 
         dispatch({ type: 'SET_ABOUT', payload: data })
       } catch (error) {
-        dispatch({ type: 'SET_ERROR', payload: 'Unable to load /about' })
+        dispatch({ type: 'SET_ERROR', payload: 'Unable to load about' })
       }
     }
 
@@ -89,13 +89,13 @@ function Info({ post }) {
       <h6 className="mb-1">{post.name}</h6>
       {post.info && <p className="my-0 py-0 fw-light">{post.info}</p>}
 
-      {post.tags?.map((button, i) => (
-        <Badge key={i} tag={button} classes={'bg-white text-dark fw-normal text-uppercase hoverable-badge'} />
+      {post.tags?.map((tag, i) => (
+        <Badge key={i} tag={tag} classes={'bg-white text-dark fw-normal text-uppercase hoverable-badge'} />
       ))}
     </div>
   )
 }
 
-const ProfileImage = ({ children }) => <img src="./profile.png" alt="" className="img-fluid  rounded" />
+const ProfileImage = ({ children }) => <img src="./profile.png" alt="sameer_image" className="img-fluid  rounded" />
 
 export { AboutTabs }
